feat: creating a cache on job fail
This removes the responsibility of knowing when a cache should be created in relation to the result of the tasks that have been run from the job from the current github action. This is useful in situations where a monorepo is being used because even if part of the job fails, the cache will still be created, which is especially true in instances where nx is used because it already only caches on successful runs. For example, a job with 3 cacheable tasks that succeed should not be penalized because the last task fails...